pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyshotter"
description = "PyShotter: Smart, annotated, and shareable screenshots for Python."
readme = "README.md"
requires-python = ">= 3.9"
authors = [
  { name = "Abdoullah Ndao", email="abdoullahaljersi@gmail.com" },
]
maintainers = [
  { name = "Abdoullah Ndao", email="abdoullahaljersi@gmail.com" },
]
license = { file = "LICENSE.txt" }
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Windows",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: Unix",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
  "Topic :: Software Development :: Libraries",
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Text Processing :: Linguistic",
]
keywords = [
    "screenshot",
    "screen capture",
    "screen recording",
    "annotation",
    "code beautification",
    "sharing",
    "smart detection",
    "ocr",
    "text extraction",
    "redaction",
    "privacy",
    "face blur",
    "panorama",
    "change detection",
    "hotkeys",
    "history",
    "gui",
    "python",
    "cross-platform",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/utachicodes/pyshotter"
Documentation = "https://github.com/utachicodes/pyshotter#readme"
Changelog = "https://github.com/utachicodes/pyshotter/blob/main/CHANGELOG.md"
Source = "https://github.com/utachicodes/pyshotter"
Tracker = "https://github.com/utachicodes/pyshotter/issues"

[project.scripts]
pyshotter = "pyshotter.__main__:main"
pyshotter-gui = "pyshotter.gui:main_gui"

[project.optional-dependencies]
# Existing groups (enhanced)
ocr = [
  "pytesseract>=0.3.10",
  "opencv-python>=4.8.0",
]
annotation = [
  "opencv-python>=4.8.0",
  "pillow>=10.0.0",
  "numpy>=1.24.0",
]
sharing = [
  "pyperclip>=1.8.0",
  "requests>=2.31.0",
]
cloud = [
  "requests>=2.31.0",
  "dropbox>=11.36.0",
]

# New v1.1 feature groups
recording = [
  "imageio>=2.31.0",
  "imageio-ffmpeg>=0.4.9",
  "pillow>=10.0.0",
  "numpy>=1.24.0",
]
gui = [
  "pystray>=0.19.4",
  "pillow>=10.0.0",
  "pynput>=1.7.6; sys_platform == 'darwin'",
  "keyboard>=0.13.5; sys_platform == 'win32'",
]
ai = [
  "opencv-python>=4.8.0",
  "numpy>=1.24.0",
  "pillow>=10.0.0",
]
config = [
  "pyyaml>=6.0",
  "pydantic>=2.0.0",
]
cli = [
  "rich>=13.0.0",
]

# Meta groups
full = [
  "pyshotter[ocr,annotation,recording,gui,ai,config,cli,sharing]",
]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.11.1",
  "pytest-benchmark>=4.0.0",
  "ruff>=0.1.0",
  "mypy>=1.5.0",
  "bandit>=1.7.5",
  "safety>=2.3.0",
]
tests = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pytest-xvfb>=3.0.0",
]
docs = [
  "sphinx>=7.0.0",
  "sphinx-rtd-theme>=1.3.0",
]

[tool.hatch.version]
path = "src/pyshotter/__init__.py"

[tool.hatch.build]
skip-excluded-dirs = true

[tool.hatch.build.targets.sdist]
only-include = [
  "CHANGELOG.md",
  "CHANGES.md",
  "CONTRIBUTORS.md",
  "docs/source",
  "src",
]

[tool.hatch.build.targets.wheel]
packages = [
  "src/pyshotter",
]
